Fixing misc compiler warnings in windows
git-svn-id: file:///Users/joaander/devel/hoomd-git-convert/repos/trunk@1420 fa922fa7-2fde-0310-acd8-f43f465a7996

+++ src/updaters/BD_NVTUpdater.cc
@@ -105,7 +105,7 @@ void BD_NVTUpdater::addStochasticBath()
 			this->addForceCompute(m_bdfc);	
 			}
 
-		m_bath_index = m_forces.size() - 1;	
+		m_bath_index = (unsigned int)m_forces.size() - 1;	
 		
 		boost::shared_ptr<StochasticForceCompute> stochastic_force(boost::shared_dynamic_cast<StochasticForceCompute>(m_forces[m_bath_index]));	
 		assert(stochastic_force);
